Various updates and fixes

+++ Backend/WarehouseAPI/WarehouseAPI/Models/DTOs/AddGoodsDto.cs
@@ -11,9 +11,9 @@ namespace WarehouseAPI.Models.DTOs
         [Required]
         public string? Name { get; set; }
 
-        public float? Vat { get; set; }
+        public decimal? Vat { get; set; }
 
-        public float? Sprice { get; set; }
+        public decimal? Sprice { get; set; }
 
         public float? MinStock { get; set; }
